# compiled-swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: CMP-API
servers:
  - url: 'https://cmp-dev.transerve.com/api'
  - url: ' http://localhost:3000'
host: 'localhost:4000'
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /project/city/code/add:
    post:
      security:
        - token: []
      summary: add city code
      tags:
        - PROJECT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - citycode
                - cityname
              properties:
                citycode:
                  type: string
                  example: CITIIS-OR-BBR
                cityname:
                  type: string
                  example: vizag
                description:
                  type: string
                  example: project for develop schools
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: object
                    properties:
                      reference:
                        type: string
                        example: CITIIS-OR-BBR
                      name:
                        type: string
                        example: vizag
                      summary:
                        type: string
                        example: project for develop schools
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/city/code/edit/{id}':
    post:
      security:
        - token: []
      summary: edit city code
      tags:
        - PROJECT
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: project id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - citycode
                - cityname
              properties:
                citycode:
                  type: string
                  example: CITIIS-OR-BBR
                cityname:
                  type: string
                  example: vizag
                description:
                  type: string
                  example: project for develop schools
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: object
                    properties:
                      reference:
                        type: string
                        example: CITIIS-OR-BBR
                      name:
                        type: string
                        example: vizag
                      summary:
                        type: string
                        example: project for develop schoolss
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
...
  /project/list:
    get:
      security:
        - token: []
      summary: get project list
      tags:
        - PROJECT
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: project_id
                        name:
                          type: string
                          example: howrah
                        city:
                          type: string
                          example: agarthala
                        maturationEndDate:
                          type: string
                          example: 22-1-2001
                        theme:
                          type: string
                          example: theme1
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/{id}':
    get:
      security:
        - token: []
      summary: get project list
      tags:
        - PROJECT
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: project id
      responses:
        '200':
          $ref: './schema/#responses/project'
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/{id}/task/list':
    get:
      security:
        - token: []
      summary: get project list
      tags:
        - PROJECT
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: project id
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  list:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: Taskname
                        startDate:
                          type: string
                          example: 22-1-2001
                        endDate:
                          type: string
                          example: 23-2-2019
                        access:
                          type: array
                          items:
                            type: object
                            properties:
                              type:
                                type: string
                                example: owner
                              role:
                                type: string
                                example: mentor
                              status:
                                type: string
                                example: PENDING
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
...
  /project/tag/add:
    post:
      security:
        - token: []
      summary: add tag
      tags:
        - TAG
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tag
              properties:
                tag:
                  type: string
                  example: document
                description:
                  type: string
                  example: document for vizag project
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/tag/edit/{id}':
    post:
      security:
        - token: []
      summary: edit tag
      tags:
        - TAG
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: tag _id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag:
                  type: string
                  example: document
                description:
                  type: string
                  example: document for vizag project
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/tag/status/{id}':
    put:
      security:
        - token: []
      summary: active or inactive tag
      tags:
        - TAG
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: tag _id
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
...
  /project/theme/add:
    post:
      security:
        - token: []
      summary: add theme
      tags:
        - THEME
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - theme
              properties:
                theme:
                  type: string
                  example: dams
                description:
                  type: string
                  example: dam for vizag
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/theme/edit/{id}':
    post:
      security:
        - token: []
      summary: edit theme
      tags:
        - THEME
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: theme _id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                theme:
                  type: string
                  example: Dam
                description:
                  type: string
                  example: Dam for vizag
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  '/project/theme/status/{id}':
    put:
      security:
        - token: []
      summary: active or inactive theme
      tags:
        - THEME
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: theme _id
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
  /project/theme/list:
    get:
      security:
        - token: []
      summary: get theme list
      tags:
        - THEME
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  'status:':
                    type: boolean
                    example: true
                  data:
                    type: object
        '400':
          $ref: '#/definitions/ErrorResponse/BadRequest'
...
components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      bearerFormat: JWT
...